Rename effect.ts to use-effect-once.ts
