+++ apps/client/src/routes/(onboarding)/onboarding.tsx
@@ -3,7 +3,7 @@ import { createFileRoute, redirect } from "@tanstack/react-router"
 
 import { onboardingMutationOptions } from "@/features/onboarding/api/options"
 import { OnboardingAnimation } from "@/features/onboarding/ui/multi-step-loader"
-import { useEffectOnce } from "@/hooks/effect"
+import { useEffectOnce } from "@/hooks/use-effect-once"
 import { Page } from "@/libs/ui/page"
 import { isAuthenticated, isOnboarded } from "@/utils/auth"
